requirements.txt
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
tep==0.8.2
channels==3.0.3
django-apscheduler==0.5.2
drf_yasg
python-jenkins